requirements2.txt
absl-py==2.3.0
accelerate==1.7.0
addict==2.4.0
annotated-types==0.7.0
anthropic==0.55.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
cached-property==1.5.2
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
classproperties==0.2.0
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.1.0
datafiles==2.3.3
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.2
fsspec==2024.6.1
gitdb==4.0.12
gitman==3.5.3
GitPython==3.1.44
grpcio==1.72.1
gymnasium==1.1.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
hidapi==0.14.0.post2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hydra-core==1.3.2
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipython==8.12.3
ipywidgets==8.1.7
isaacsim==4.5.0.0
isaacsim-app==4.5.0.0
isaacsim-asset==4.5.0.0
isaacsim-benchmark==4.5.0.0
isaacsim-code-editor==4.5.0.0
isaacsim-core==4.5.0.0
isaacsim-cortex==4.5.0.0
isaacsim-example==4.5.0.0
isaacsim-extscache-kit==4.5.0.0
isaacsim-extscache-kit-sdk==4.5.0.0
isaacsim-extscache-physics==4.5.0.0
isaacsim-gui==4.5.0.0
isaacsim-kernel==4.5.0.0
isaacsim-replicator==4.5.0.0
isaacsim-rl==4.5.0.0
isaacsim-robot==4.5.0.0
isaacsim-robot-motion==4.5.0.0
isaacsim-robot-setup==4.5.0.0
isaacsim-ros1==4.5.0.0
isaacsim-ros2==4.5.0.0
isaacsim-sensor==4.5.0.0
isaacsim-storage==4.5.0.0
isaacsim-template==4.5.0.0
isaacsim-test==4.5.0.0
isaacsim-utils==4.5.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json-five==1.1.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
minilog==2.3.1
mistune==3.1.3
moviepy==2.2.1
mpmath==1.3.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
omniverse-kit==106.5.0.162521
onnx==1.16.1
open3d==0.19.0
opencv-python==4.11.0.86
orjson==3.10.18
osqp==0.6.7.post3
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.8
plotly==6.2.0
pre_commit==4.2.0
prettytable==3.3.0
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyglet==1.5.31
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qdldl==0.1.7.post5
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rerun==1.0.30
retrying==1.4.0
rpds-py==0.25.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segment-anything==1.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
sly==0.5
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensordict==0.8.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.10.2
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trimesh==4.6.11
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
universal-startfile==0.2
urllib3==2.4.0
virtualenv==20.31.2
wandb==0.20.1
warp-lang==1.7.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yarg==0.1.9
zipp==3.22.0
